requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
mysqlclient==2.2.4
pygame==2.6.0
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
